# SchoolManagement/Endpoints/StudentEndpoints.cs
using SchoolManagement.Models;
using SchoolManagement.Services;

namespace SchoolManagement.Endpoints;

public static class StudentEndpoints
{
    public static void MapStudentEndpoints(this IEndpointRouteBuilder app)
    {
        var group = app.MapGroup("/api/students")
            .WithTags("Students");

        group.MapGet("/", GetAllStudents)
            .WithName("GetAllStudents")
            .Produces<IEnumerable<StudentDto>>();

        group.MapGet("/{id}", GetStudentById)
            .WithName("GetStudentById")
            .Produces<StudentDto>()
            .Produces(StatusCodes.Status404NotFound);

        group.MapPost("/", CreateStudent)
            .WithName("CreateStudent")
            .Produces<StudentDto>(StatusCodes.Status201Created)
            .Produces(StatusCodes.Status400BadRequest)
            .Produces(StatusCodes.Status409Conflict);

        group.MapPut("/{id}", UpdateStudent)
            .WithName("UpdateStudent")
            .Produces<StudentDto>()
            .Produces(StatusCodes.Status400BadRequest)
            .Produces(StatusCodes.Status404NotFound);

        group.MapPatch("/{id}", PatchStudent)
            .WithName("PatchStudent")
            .WithDescription("Partially update a student. Use this to assign/unassign students to/from classes.")
            .Produces<StudentDto>()
            .Produces(StatusCodes.Status400BadRequest)
            .Produces(StatusCodes.Status404NotFound);

        group.MapDelete("/{id}", DeleteStudent)
            .WithName("DeleteStudent")
            .Produces<object>()
            .Produces(StatusCodes.Status404NotFound);
    }

    private static async Task<IResult> GetAllStudents(IStudentService studentService)
    {
        var students = await studentService.GetAllStudentsAsync();
        return Results.Ok(students);
    }

    private static async Task<IResult> GetStudentById(string id, IStudentService studentService)
    {
        var result = await studentService.GetStudentByIdAsync(id);
        return result.ToHttpResult();
    }

    private static async Task<IResult> CreateStudent(CreateStudentDto dto, IStudentService studentService)
    {
        var result = await studentService.CreateStudentAsync(dto);
        return result.Status == ServiceResultStatus.Created
            ? Results.Created($"/api/students/{result.Data?.StudentId}", result.Data)
            : result.ToHttpResult();
    }

    private static async Task<IResult> UpdateStudent(string id, UpdateStudentDto dto, IStudentService studentService)
    {
        var result = await studentService.UpdateStudentAsync(id, dto);
        return result.ToHttpResult();
    }

    private static async Task<IResult> PatchStudent(string id, PatchStudentDto dto, IStudentService studentService)
    {
        var result = await studentService.PatchStudentAsync(id, dto);
        return result.ToHttpResult();
    }

    private static async Task<IResult> DeleteStudent(string id, IStudentService studentService)
    {
        var result = await studentService.DeleteStudentAsync(id);
        return result.ToHttpResult(data => new { message = data });
    }
}