# SchoolManagement/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "School Management API",
    "description": "API for managing students and school classes",
    "version": "1.0"
  },
  "paths": {
    "/api/students": {
      "get": {
        "tags": ["Students"],
        "summary": "Get all students",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": ["Students"],
        "summary": "Create a new student",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "409": {
            "description": "Conflict - Student ID already exists"
          }
        }
      }
    },
    "/api/students/{id}": {
      "get": {
        "tags": ["Students"],
        "summary": "Get student by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": ["Students"],
        "summary": "Update an existing student",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": ["Students"],
        "summary": "Delete a student",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/classes": {
      "get": {
        "tags": ["Classes"],
        "summary": "Get all school classes",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": ["Classes"],
        "summary": "Create a new school class",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSchoolClassDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/classes/{id}": {
      "get": {
        "tags": ["Classes"],
        "summary": "Get school class by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": ["Classes"],
        "summary": "Update an existing school class",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSchoolClassDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": ["Classes"],
        "summary": "Delete a school class",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/classes/{classId}/students": {
      "post": {
        "tags": ["Classes"],
        "summary": "Add student to class",
        "parameters": [
          {
            "name": "classId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddStudentToClassDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request - Class full or student already in class"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/classes/{classId}/students/{studentId}": {
      "delete": {
        "tags": ["Classes"],
        "summary": "Remove student from class",
        "parameters": [
          {
            "name": "classId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateStudentDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "surname": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date-time"
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          }
        },
        "required": ["studentId", "name", "surname", "dateOfBirth"]
      },
      "UpdateStudentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "surname": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date-time"
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          }
        },
        "required": ["name", "surname", "dateOfBirth"]
      },
      "CreateSchoolClassDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "leadingTeacher": {
            "type": "string"
          }
        },
        "required": ["name", "leadingTeacher"]
      },
      "UpdateSchoolClassDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "leadingTeacher": {
            "type": "string"
          }
        },
        "required": ["name", "leadingTeacher"]
      },
      "AddStudentToClassDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string"
          }
        },
        "required": ["studentId"]
      }
    }
  }
}